+++ library/ui/src/main/java/com/upax/exoplayer2/ui/StyledPlayerView.java
@@ -117,7 +117,7 @@ import org.checkerframework.checker.nullness.qual.RequiresNonNull;
  *         <li>Corresponding method: {@link #setControllerHideDuringAds(boolean)}
  *         <li>Default: {@code true}
  *       </ul>
- *   <li><b>{@code show_buffering}</b> - Whether the buffering spinner is displayed when the player
+ *   <li><b>{@code show_buffering_change}</b> - Whether the buffering spinner is displayed when the player
  *       is buffering. Valid values are {@code never}, {@code when_playing} and {@code always}.
  *       <ul>
  *         <li>Corresponding method: {@link #setShowBuffering(int)}
@@ -384,7 +384,7 @@ public class StyledPlayerView extends FrameLayout implements AdViewProvider {
             a.getBoolean(R.styleable.StyledPlayerView_hide_on_touch, controllerHideOnTouch);
         controllerAutoShow =
             a.getBoolean(R.styleable.StyledPlayerView_auto_show, controllerAutoShow);
-        showBuffering = a.getInteger(R.styleable.StyledPlayerView_show_buffering, showBuffering);
+        showBuffering = a.getInteger(R.styleable.StyledPlayerView_show_buffering_change, showBuffering);
         keepContentOnPlayerReset =
             a.getBoolean(
                 R.styleable.StyledPlayerView_keep_content_on_player_reset,