+++ RELEASENOTES.md
@@ -2490,7 +2490,7 @@ To learn more about what's new in 2.12, read the corresponding
     ([#5097](https://github.com/google/ExoPlayer/issues/5097)).
 *   Include channel count in audio capabilities check
     ([#4690](https://github.com/google/ExoPlayer/issues/4690)).
-*   Fix issue with applying the `show_buffering` attribute in `PlayerView`
+*   Fix issue with applying the `show_buffering_change` attribute in `PlayerView`
     ([#5139](https://github.com/google/ExoPlayer/issues/5139)).
 *   Fix issue where null `Metadata` was output when it failed to decode
     ([#5149](https://github.com/google/ExoPlayer/issues/5149)).